Add files via upload
so need to make this stuff work together now i guees and stuff
